Improve UI (#37)
* Fix service list filtering on page reload.

When filtering the service list, a query parameter is appended to the URL,
but unused when reloaded.

* Factor similar conditions in `ConsulService`.

* Fix crash when iterating on an empty list.

`for ... in` will iterate on internal properties (like `length`), and this
will crash.

* Use `fetch()` instead of jQuery.

* Simplify service insertion in list.

* Make service loading non-blocking.

* Fix typo.

* Hide the horizontal scrollbar in the Nodes page.

* Debounce search fields.

Search will start 250 ms after the last key up. This is an easy way to avoid
blocking the browser when typing.